feat:갤러리 권한 기능
