+++ JunYoung/week3/package.json
@@ -13,6 +13,7 @@
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
     "expo": "~51.0.14",
+    "expo-image-picker": "~15.0.5",
     "expo-linear-gradient": "~13.0.2",
     "expo-splash-screen": "~0.27.5",
     "expo-status-bar": "~1.12.1",
